# Computed initial penalty values