--- a/examples/using_pyvrp_components.ipynb
+++ b/examples/using_pyvrp_components.ipynb
@@ -241,7 +241,7 @@
    "source": [
     "from pyvrp import PenaltyManager\n",
     "\n",
-    "pen_manager = PenaltyManager.initial_from_data(INSTANCE)\n",
+    "pen_manager = PenaltyManager.init_from_data(INSTANCE)\n",
     "cost_evaluator = pen_manager.cost_evaluator()"
    ]
   },
@@ -436,7 +436,7 @@
    "source": [
     "def solve(stop, seed):\n",
     "    rng = RandomNumberGenerator(seed=seed)\n",
-    "    pm = PenaltyManager.initial_from_data(INSTANCE)\n",
+    "    pm = PenaltyManager.init_from_data(INSTANCE)\n",
     "    pop = Population(broken_pairs_distance)\n",
     "\n",
     "    neighbours = compute_neighbours(INSTANCE)\n",