--- a/tests/test_PenaltyManager.py
+++ b/tests/test_PenaltyManager.py
@@ -7,9 +7,6 @@ from pyvrp.exceptions import PenaltyBoundWarning
 
 @pytest.mark.parametrize(
     (
-        "init_load_penalty",
-        "init_tw_penalty",
-        "init_dist_penalty",
         "repair_booster",
         "solutions_between_updates",
         "penalty_increase",
@@ -17,19 +14,16 @@ from pyvrp.exceptions import PenaltyBoundWarning
         "target_feasible",
     ),
     [
-        (1, 1, 1, 1, 0, -1.0, 0.5, 0.5),  # -1 penalty increase
-        (1, 1, 1, 1, 0, 0.5, 0.5, 0.5),  # 0.5 penalty increase
-        (1, 1, 1, 1, 0, 1.5, -1, 0.5),  # -1 penalty decrease
-        (1, 1, 1, 1, 0, 1.5, 2, 0.5),  # 2 penalty decrease
-        (1, 1, 1, 1, 0, 1, 1, -1),  # -1 target feasible
-        (1, 1, 1, 1, 0, 1, 1, 2),  # 2 target feasible
-        (1, 1, 1, 0, 0, 1, 1, 1),  # 0 repair booster
+        (1, 0, -1.0, 0.5, 0.5),  # -1 penalty increase
+        (1, 0, 0.5, 0.5, 0.5),  # 0.5 penalty increase
+        (1, 0, 1.5, -1, 0.5),  # -1 penalty decrease
+        (1, 0, 1.5, 2, 0.5),  # 2 penalty decrease
+        (1, 0, 1, 1, -1),  # -1 target feasible
+        (1, 0, 1, 1, 2),  # 2 target feasible
+        (0, 0, 1, 1, 1),  # 0 repair booster
     ],
 )
 def test_constructor_throws_when_arguments_invalid(
-    init_load_penalty: int,
-    init_tw_penalty: int,
-    init_dist_penalty: int,
     repair_booster: int,
     solutions_between_updates: int,
     penalty_increase: float,
@@ -41,9 +35,6 @@ def test_constructor_throws_when_arguments_invalid(
     """
     with assert_raises(ValueError):
         PenaltyParams(
-            init_load_penalty,
-            init_tw_penalty,
-            init_dist_penalty,
             repair_booster,
             solutions_between_updates,
             penalty_increase,
@@ -57,8 +48,8 @@ def test_repair_booster():
     Tests that the booster evaluator returns a cost evaluator object that
     penalises constraint violations much more severely.
     """
-    params = PenaltyParams(1, 1, 1, 5, 1, 1, 1, 1)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(5, 1, 1, 1, 1)
+    pm = PenaltyManager(params, initial_penalties=(1, 1, 1))
 
     cost_evaluator = pm.cost_evaluator()
 
@@ -85,8 +76,8 @@ def test_load_penalty_update_increase(ok_small):
     infeasible solutions have been generated since the last update.
     """
     num_registrations = 4
-    params = PenaltyParams(1, 1, 1, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(1, 1, 1))
 
     # Within bandwidth, so penalty should not change.
     assert_equal(pm.cost_evaluator().load_penalty(2, 1), 1)
@@ -101,18 +92,17 @@ def test_load_penalty_update_increase(ok_small):
         pm.register(sol)
     assert_equal(pm.cost_evaluator().load_penalty(2, 1), 1)
 
-    # Below targetFeasible, so should increase the capacityPenalty by +1
-    # (normally to 1.1 due to penaltyIncrease, but we should not end up at the
-    # same int).
+    # Below targetFeasible, so should increase the loadPenalty by +1 (normally
+    # to 1.1 due to penaltyIncrease, but we should not end up at the same int).
     for sol in [infeas] * num_registrations:
         pm.register(sol)
     assert_equal(pm.cost_evaluator().load_penalty(2, 1), 2)
 
-    # Now we start from a much bigger initial capacityPenalty. Here we want the
+    # Now we start from a much bigger initial loadPenalty. Here we want the
     # penalty to increase by 10% due to penaltyIncrease = 1.1, and +1 due to
     # double -> int.
-    params = PenaltyParams(100, 1, 1, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(100, 1, 1))
 
     assert_equal(pm.cost_evaluator().load_penalty(2, 1), 100)
     for sol in [infeas] * num_registrations:
@@ -126,8 +116,8 @@ def test_load_penalty_update_decrease(ok_small):
     load feasible solutions have been generated since the last update.
     """
     num_registrations = 4
-    params = PenaltyParams(4, 1, 1, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(4, 1, 1))
 
     feas = Solution(ok_small, [[1, 2]])
     infeas = Solution(ok_small, [[1, 2, 3]])
@@ -141,17 +131,17 @@ def test_load_penalty_update_decrease(ok_small):
         pm.register(sol)
     assert_equal(pm.cost_evaluator().load_penalty(2, 1), 4)
 
-    # Above targetFeasible, so should decrease the capacityPenalty to 90%, and
-    # -1 from the bounds check. So 0.9 * 4 = 3.6, 3.6 - 1 = 2.6, (int) 2.6 = 2
+    # Above targetFeasible, so should decrease the loadPenalty to 90%, and -1
+    # from the bounds check. So 0.9 * 4 = 3.6, 3.6 - 1 = 2.6, (int) 2.6 = 2
     for sol in [feas] * num_registrations:
         pm.register(sol)
     assert_equal(pm.cost_evaluator().load_penalty(2, 1), 2)
 
-    # Now we start from a much bigger initial capacityPenalty. Here we want the
+    # Now we start from a much bigger initial loadPenalty. Here we want the
     # penalty to decrease by 10% due to penaltyDecrease = 0.9, and -1 due to
     # double -> int.
-    params = PenaltyParams(100, 1, 1, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(100, 1, 1))
 
     assert_equal(pm.cost_evaluator().load_penalty(2, 1), 100)
     for sol in [feas] * num_registrations:
@@ -159,8 +149,8 @@ def test_load_penalty_update_decrease(ok_small):
     assert_equal(pm.cost_evaluator().load_penalty(2, 1), 89)
 
     # Test that the penalty cannot decrease beyond 1, its minimum value.
-    params = PenaltyParams(1, 1, 1, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(1, 1, 1))
 
     assert_equal(pm.cost_evaluator().load_penalty(2, 1), 1)
     for sol in [feas] * num_registrations:
@@ -174,8 +164,8 @@ def test_time_warp_penalty_update_increase(ok_small):
     window infeasible solutions have been generated since the last update.
     """
     num_registrations = 4
-    params = PenaltyParams(1, 1, 1, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(1, 1, 1))
 
     feas = Solution(ok_small, [[1, 2]])
     infeas = Solution(ok_small, [[1, 2, 3]])
@@ -189,18 +179,17 @@ def test_time_warp_penalty_update_increase(ok_small):
         pm.register(sol)
     assert_equal(pm.cost_evaluator().tw_penalty(1), 1)
 
-    # Below targetFeasible, so should increase the tw penalty by +1
-    # (normally to 1.1 due to penaltyIncrease, but we should not end up at the
-    # same int).
+    # Below targetFeasible, so should increase the tw penalty by +1 (normally
+    # to 1.1 due to penaltyIncrease, but we should not end up at the same int).
     for sol in [infeas] * num_registrations:
         pm.register(sol)
     assert_equal(pm.cost_evaluator().tw_penalty(1), 2)
 
-    # Now we start from a much bigger initial tw penalty. Here we want
-    # the penalty to increase by 10% due to penaltyIncrease = 1.1, and +1 due
+    # Now we start from a much bigger initial tw penalty. Here we want the
+    # penalty to increase by 10% due to penaltyIncrease = 1.1, and +1 due
     # to double -> int.
-    params = PenaltyParams(1, 100, 1, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(1, 100, 1))
 
     assert_equal(pm.cost_evaluator().tw_penalty(1), 100)
     for sol in [infeas] * num_registrations:
@@ -215,8 +204,8 @@ def test_time_warp_penalty_update_decrease(ok_small):
     update.
     """
     num_registrations = 4
-    params = PenaltyParams(1, 4, 1, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(1, 4, 1))
 
     feas = Solution(ok_small, [[1, 2]])
     infeas = Solution(ok_small, [[1, 2, 3]])
@@ -230,17 +219,17 @@ def test_time_warp_penalty_update_decrease(ok_small):
         pm.register(sol)
     assert_equal(pm.cost_evaluator().tw_penalty(1), 4)
 
-    # Above targetFeasible, so should decrease the timeWarpPenalty to 90%, and
-    # -1 from the bounds check. So 0.9 * 4 = 3.6, 3.6 - 1 = 2.6, (int) 2.6 = 2
+    # Above targetFeasible, so should decrease the twCapacity to 90%, and -1
+    # from the bounds check. So 0.9 * 4 = 3.6, 3.6 - 1 = 2.6, (int) 2.6 = 2.
     for sol in [feas] * num_registrations:
         pm.register(sol)
     assert_equal(pm.cost_evaluator().tw_penalty(1), 2)
 
-    # Now we start from a much bigger initial timeWarpCapacity. Here we want
-    # the penalty to decrease by 10% due to penaltyDecrease = 0.9, and -1 due
+    # Now we start from a much bigger initial twCapacity. Here we want the
+    # penalty to decrease by 10% due to penaltyDecrease = 0.9, and -1 due
     # to double -> int.
-    params = PenaltyParams(1, 100, 1, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(1, 100, 1))
 
     assert_equal(pm.cost_evaluator().tw_penalty(1), 100)
     for sol in [feas] * num_registrations:
@@ -248,8 +237,8 @@ def test_time_warp_penalty_update_decrease(ok_small):
     assert_equal(pm.cost_evaluator().tw_penalty(1), 89)
 
     # Test that the penalty cannot decrease beyond 1, its minimum value.
-    params = PenaltyParams(1, 1, 1, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(1, 1, 1))
 
     assert_equal(pm.cost_evaluator().tw_penalty(1), 1)
     for sol in [feas] * num_registrations:
@@ -266,8 +255,8 @@ def test_does_not_update_penalties_before_sufficient_registrations(ok_small):
     data = ok_small.replace(vehicle_types=[vehicle_type])
 
     num_registrations = 4
-    params = PenaltyParams(4, 4, 4, 1, num_registrations, 1.1, 0.9, 0.5)
-    pm = PenaltyManager(params)
+    params = PenaltyParams(1, num_registrations, 1.1, 0.9, 0.5)
+    pm = PenaltyManager(params, initial_penalties=(4, 4, 4))
 
     feas = Solution(data, [[1, 2], [3, 4]])
     infeas = Solution(data, [[1, 2, 3, 4]])
@@ -303,12 +292,11 @@ def test_max_min_penalty(ok_small):
     when updating their values.
     """
     params = PenaltyParams(
-        init_time_warp_penalty=PenaltyManager.MAX_PENALTY,
         solutions_between_updates=1,
         penalty_decrease=0,
         penalty_increase=2,
     )
-    pm = PenaltyManager(params)
+    pm = PenaltyManager(params, (20, PenaltyManager.MAX_PENALTY, 6))
 
     # Initial penalty is MAX_PENALTY, so one unit of time warp should be
     # penalised by that value.
@@ -338,11 +326,9 @@ def test_warns_max_penalty_value(ok_small):
     Tests that a penalty parameter clipped to MAX_PENALTY raises a warning.
     This typically indicates a data issue that PyVRP is struggling with.
     """
-    params = PenaltyParams(
-        init_time_warp_penalty=PenaltyManager.MAX_PENALTY,
-        solutions_between_updates=1,
-    )
-    pm = PenaltyManager(params)
+    params = PenaltyParams(solutions_between_updates=1)
+    initial = (1, PenaltyManager.MAX_PENALTY, 1)
+    pm = PenaltyManager(params, initial_penalties=initial)
 
     infeas = Solution(ok_small, [[1, 2, 3, 4]])
     assert_(infeas.has_time_warp())