@N-Wouda 2024-06-11
@leonlan could you have a look at this? And, should we benchmark this?