@N-Wouda 2024-06-12
My new default penalty parameters are `[13040    11     1]`, which is a lot better than what I previously got.